+++ lib/hardware/behavior.ex
@@ -1,0 +1,7 @@
+defmodule Replicate.Hardware.Behaviour do
+  @moduledoc """
+  Documentation for the Hardware Behaviour
+  """
+  alias Replicate.Hardware.Hardware
+  @callback list :: list(Hardware.t())
+end