+++ test/replicate_test.exs
@@ -3,10 +3,12 @@ defmodule ReplicateTest do
   import Mox
   alias Replicate.Predictions.Prediction
   alias Replicate.Models.Model
+  alias Replicate.Hardware.Hardware
   doctest Replicate
   doctest Replicate.Predictions
   doctest Replicate.Models
   doctest Replicate.Deployments
+  doctest Replicate.Hardware
 
   # Make sure mocks are verified when the test exits
   setup :verify_on_exit!
@@ -128,5 +130,34 @@ defmodule ReplicateTest do
     first_batch = stream |> Enum.at(0)
 
     assert first_batch |> length() == 25
+  end
+
+  test "create a model" do
+    {:ok, model} =
+      Replicate.Models.create(
+        owner: "replicate",
+        name: "babadook",
+        visibility: "public",
+        hardware: "gpu-a40-large"
+      )
+
+    assert model.owner == "replicate"
+    assert model.name == "babadook"
+    assert model.visibility == "public"
+
+    {:error, message} =
+      Replicate.Models.create(
+        owner: "replicate",
+        name: "babadook"
+      )
+
+    assert message == "A required parameter (owner/name/visiblity/hardware) is missing"
+  end
+
+  test "list hardware" do
+    hardware = Replicate.Hardware.list()
+
+    assert hardware |> length() == 4
+    assert hardware |> Enum.at(0) == %Hardware{name: "CPU", sku: "cpu"}
   end
 end