+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-Nothing to see here, yet!
+## [2023-11-06]
+
+- Added `Replicate.Models.create/4` function to create a model for a user or organization with a given name, visibility, and hardware SKU.
+- Added `Replicate.Hardware.list/0` function to list available hardware SKUs.